scripts to update sqlite db
